-Tipos de dados das coordenadas de quadros e tamanho de quadros foram alterados de [int]para [double].

-Refatoração de todos os endpoints do sistema, bem como reestruturação da interface ICanvasService e suas implementações em CanvasService.

-As funções do ICanvasServices retornam um booleano referente ao sucesso ou fracasso da execução da requisição, atribuindo através do [out variavel] o objeto resultante do processamento, resultando em uma padronização no tratamento de requests por parte dos endpoints.

+++ server/src/PushIt/DTOs_Contracts/CreateQuadroRequest.cs
@@ -1,10 +1,10 @@
 public record CreateQuadroRequest
 (
     string id,
-    int x,
-    int y,
-    int width,
-    int height,
+    double x,
+    double y,
+    double width,
+    double height,
     string text,
     string colour
 );
